Create model comparison

DALEX is designed to work with various black-box models like tree ensembles, linear models, neural networks etc. Threfore we may want to compare our models using such a usefull tool.

champion_challenger(champion, challenger, variable = NULL, type,
  control = rpart::rpart.control(maxdepth = 4, minbucket = 27, cp =
  0.0082212), ...)

Arguments

champion
- explainer of model that is supposed to be challenged

challenger
- explainer or list of explainers of models that are supposed to beat champion

variable
- variable that is going to be used in "prediction versus varaible" comparison

type
- classification or regression

control
- rpart.control object that will be passed to rpart tree

...
- other arguments that will ba passed to rmarkdown::render()

Value

html raport

Examples

if(DALEXtra:::is_conda()) {
library("DALEXtra")
titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
titanic_train <- read.csv(system.file("extdata", "titanic_train.csv", package = "DALEXtra"))
library("mlr")
task <- mlr::makeClassifTask(
id = "R",
data = titanic_train,
target = "survived"
)
learner <- mlr::makeLearner(
  "classif.gbm",
  par.vals = list(
    distribution = "bernoulli",
    n.trees = 500,
    interaction.depth = 4,
    n.minobsinnode = 12,
    shrinkage = 0.001,
    bag.fraction = 0.5,
    train.fraction = 1
  ),
  predict.type = "prob"
)
gbm <- mlr::train(learner, task)
explainer_mlr <- explain_mlr(gbm, titanic_test[,1:17], titanic_test[,18])

h2o::h2o.init()
h2o::h2o.no_progress()
titanic_h2o <- h2o::as.h2o(titanic_train)
titanic_h2o["survived"] <- h2o::as.factor(titanic_h2o["survived"])
titanic_test_h2o <- h2o::as.h2o(titanic_test)
model <- h2o::h2o.gbm(
training_frame = titanic_h2o,
y = "survived",
distribution = "bernoulli",
ntrees = 500,
max_depth = 4,
min_rows =  12,
learn_rate = 0.001
)
explainer_h2o <- explain_h2o(model, titanic_test[,1:17], titanic_test[,18])


# Explainer build (Keep in mind that 18th column is target)
titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
# Keep in mind that when pickle is being built and loaded,
# not only Python version but libraries versions has to match aswell
explainer_scikit <- explain_scikitlearn(system.file("extdata",
                                                    "scikitlearn.pkl",
                                                    package = "DALEXtra"),
                                        yml = system.file("extdata",
                                                          "scikitlearn_unix.yml",
                                                          package = "DALEXtra"),
                                        data = titanic_test[,1:17],
                                        y = titanic_test$survived)

champion_challenger(explainer_mlr, list(explainer_scikit, explainer_h2o), type = "classification",
                    variable = c("fare", "age"))
} else {
  print('Conda is required.')
}

#> Loading required package: ParamHelpers

#> Preparation of a new explainer is initiated
#>   -> model label       :  WrappedModel  (default)
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  yhat.WrappedModel  will be used (default)
#>   -> predicted values  :  numerical, min =  0.2693414 , mean =  0.3319247 , max =  0.5640745  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.5640745 , mean =  -0.02085597 , max =  0.7306586  
#> A new explainer has been created!
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 minutes 54 seconds 
#>     H2O cluster timezone:       Europe/Belgrade 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.24.0.5 
#>     H2O cluster version age:    2 months and 1 day  
#>     H2O cluster name:           H2O_started_from_R_szyna_czg417 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   1.97 GB 
#>     H2O cluster total cores:    4 
#>     H2O cluster allowed cores:  4 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.6.1 (2019-07-05) 
#> 
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 minutes 59 seconds 
#>     H2O cluster timezone:       Europe/Belgrade 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.24.0.5 
#>     H2O cluster version age:    2 months and 1 day  
#>     H2O cluster name:           H2O_started_from_R_szyna_czg417 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   1.97 GB 
#>     H2O cluster total cores:    4 
#>     H2O cluster allowed cores:  4 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.6.1 (2019-07-05) 
#> 
#> Preparation of a new explainer is initiated
#>   -> model label       :  H2OBinomialModel  (default)
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  yhat.H2OBinomialModel  will be used (default)
#>   -> predicted values  :  numerical, min =  0.1994385 , mean =  0.3299534 , max =  0.5876636  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.5847857 , mean =  -0.01888466 , max =  0.8005615  
#> A new explainer has been created!

#> There already exists environment named the same as it is specified in .yml file - myenv. It will be used

#> Preparation of a new explainer is initiated
#>   -> model label       :  scikitlearn_model  (default)
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  yhat.scikitlearn_model  will be used (default)
#>   -> predicted values  :  numerical, min =  0.02086126 , mean =  0.288584 , max =  0.9119996  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.8669431 , mean =  0.02248468 , max =  0.9791387  
#> A new explainer has been created!

#> 
#> 
#> processing file: ChampionChallenger_classif.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |...                                                              |   4%
#>    inline R code fragments
#> 
#>   |                                                                         |......                                                           |   9%
#> label: setup (with options) 
#> List of 3
#>  $ include: logi FALSE
#>  $ message: logi FALSE
#>  $ warning: logi TRUE
#> 
#>   |                                                                         |........                                                         |  13%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........                                                      |  17%
#> label: unnamed-chunk-1

#> 
#> Attaching package: 'auditor'

#> The following objects are masked from 'package:DALEX':
#> 
#>     model_performance, theme_drwhy

#>   |                                                                         |..............                                                   |  22%
#>   ordinary text without R code
#> 
#>   |                                                                         |.................                                                |  26%
#> label: unnamed-chunk-2

#>   |                                                                         |....................                                             |  30%
#>   ordinary text without R code
#> 
#>   |                                                                         |.......................                                          |  35%
#> label: unnamed-chunk-3
#>   |                                                                         |.........................                                        |  39%
#>   ordinary text without R code
#> 
#>   |                                                                         |............................                                     |  43%
#> label: unnamed-chunk-4
#>   |                                                                         |...............................                                  |  48%
#>   ordinary text without R code
#> 
#>   |                                                                         |..................................                               |  52%
#> label: unnamed-chunk-5

#>   |                                                                         |.....................................                            |  57%
#>   ordinary text without R code
#> 
#>   |                                                                         |........................................                         |  61%
#> label: unnamed-chunk-6

#>   |                                                                         |..........................................                       |  65%
#>   ordinary text without R code
#> 
#>   |                                                                         |.............................................                    |  70%
#> label: unnamed-chunk-7

#>   |                                                                         |................................................                 |  74%
#>   ordinary text without R code
#> 
#>   |                                                                         |...................................................              |  78%
#> label: unnamed-chunk-8

#>   |                                                                         |......................................................           |  83%
#>   ordinary text without R code
#> 
#>   |                                                                         |.........................................................        |  87%
#> label: unnamed-chunk-9

#>   |                                                                         |...........................................................      |  91%
#>   ordinary text without R code
#> 
#>   |                                                                         |..............................................................   |  96%
#> label: unnamed-chunk-10

#>   |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: ChampionChallenger_classif.knit.md

#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS ChampionChallenger_classif.utf8.md --to html4 --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output pandoc23fc425c61da.html --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_collapsed=1 --variable toc_smooth_scroll=1 --variable toc_print=1 --template "C:\Users\szyna\Documents\R\win-library\3.6\rmarkdown\rmd\h\default.html" --no-highlight --variable highlightjs=1 --number-sections --variable "theme:bootstrap" --include-in-header "C:\Users\szyna\AppData\Local\Temp\RtmpoJ7DN0\rmarkdown-str23fc5bbd4759.html" --mathjax --variable "mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" 

#> 
#> Output created: ChampionChallenger.html