Create your conda virtual env with DALEX

Phython objects may be loaded into R. However, it requiers versions of the Python and libraries to match bewtween both machines.

create_env(yml, condaenv)

Arguments

yml
a path to the .yml file. If OS is Windows conda has to be added to the PATH first

condaenv
path to main conda folder. If OS is Unix You have to specify it with .yml file path. Using with windows, param will be omitted.

Value

Name of created virtual env.

Examples

reticulate::use_condaenv("myenv")

if(.Platform$OS.type=="unix" & !("myenv" %in% reticulate::conda_list()$name)){
  create_env(system.file("extdata", "scikitlearn_unix.yml", package = "DALEXtra"),
             condaenv = "$HOME/miniconda")
}else{
  print("Use unix for tests")
}

#> [1] "Use unix for tests"