Create explainer from your h2o model

DALEX is designed to work with various black-box models like tree ensembles, linear models, neural networks etc. Unfortunately R packages that create such models are very inconsistent. Different tools use different interfaces to train, validate and use models. One of those tools, we would like to make more accessible is H2O.

explain_h2o(model, data = NULL, y = NULL, predict_function = NULL,
  residual_function = NULL, ..., label = NULL, verbose = TRUE)

Arguments

model
object - a model to be explained

data
data.frame or matrix - data that was used for fitting. If not provided then will be extracted from the model

y
numeric vector with outputs / scores. If provided then it shall have the same size as data

predict_function
function that takes two arguments: model and new data and returns numeric vector with predictions

residual_function
function that takes three arguments: model, data and response vector y. It should return a numeric vector with model residuals for given data. If not provided, response residuals (\(y-\hat{y}\)) are calculated.

...
other parameters

label
character - the name of the model. By default it's extracted from the 'class' attribute of the model

verbose
if TRUE (default) then diagnostic messages will be printed

Value

explainer object ready to work with DALEX

Examples

titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
titanic_train <- read.csv(system.file("extdata", "titanic_train.csv", package = "DALEXtra"))
library("DALEX")

#> Welcome to DALEX (version: 0.4.5).
#> Find examples and detailed introduction at: https://pbiecek.github.io/PM_VEE/

h2o::h2o.init()

#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         51 minutes 47 seconds 
#>     H2O cluster timezone:       Europe/Belgrade 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.24.0.5 
#>     H2O cluster version age:    1 month and 5 days  
#>     H2O cluster name:           H2O_started_from_R_szyna_pkc355 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   1.95 GB 
#>     H2O cluster total cores:    4 
#>     H2O cluster allowed cores:  4 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.6.1 (2019-07-05) 
#> 

h2o::h2o.no_progress()
titanic_h2o <- h2o::as.h2o(titanic_train)
titanic_h2o["survived"] <- h2o::as.factor(titanic_h2o["survived"])
titanic_test_h2o <- h2o::as.h2o(titanic_test)
model <- h2o::h2o.gbm(
training_frame = titanic_h2o,
y = "survived",
distribution = "bernoulli",
ntrees = 5000,
max_depth = 4,
min_rows =  12,
learn_rate = 0.001
)
explain_h2o(model, titanic_test[,1:17], titanic_test[,18])

#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         52 minutes 5 seconds 
#>     H2O cluster timezone:       Europe/Belgrade 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.24.0.5 
#>     H2O cluster version age:    1 month and 5 days  
#>     H2O cluster name:           H2O_started_from_R_szyna_pkc355 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   1.94 GB 
#>     H2O cluster total cores:    4 
#>     H2O cluster allowed cores:  4 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.6.1 (2019-07-05) 
#> 
#> Preparation of a new explainer is initiated
#>   -> model label       :  H2OBinomialModel  (default)
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  predict_function 
#>   -> predicted values  :  numerical, min =  0.02596615 , mean =  0.3291981 , max =  0.9759168  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.965552 , mean =  -0.0181294 , max =  0.9738659  
#> A new explainer has been created!

#> Model label:  H2OBinomialModel 
#> Model class:  H2OBinomialModel 
#> Data head  :
#>   gender.female gender.male age class.1st class.2nd class.3rd class.deck.crew
#> 1             1           0  16         0         0         1               0
#> 2             0           1  25         0         0         1               0
#>   class.engineering.crew class.restaurant.staff class.victualling.crew
#> 1                      0                      0                      0
#> 2                      0                      0                      0
#>   embarked.Belfast embarked.Cherbourg embarked.Queenstown embarked.Southampton
#> 1                0                  0                   0                    1
#> 2                0                  0                   0                    1
#>   fare sibsp parch
#> 1 7.13     0     0
#> 2 7.13     0     0