Create explainer from your mlr model

DALEX is designed to work with various black-box models like tree ensembles, linear models, neural networks etc. Unfortunately R packages that create such models are very inconsistent. Different tools use different interfaces to train, validate and use models. One of those tools, which is one of the most popular one is mlr package. We would like to present dedicated explain function for it.

explain_mlr(model, data = NULL, y = NULL, predict_function = NULL,
  residual_function = NULL, ..., label = NULL, verbose = TRUE,
  precalculate = TRUE)

Arguments

model
object - a model to be explained

data
data.frame or matrix - data that was used for fitting. If not provided then will be extracted from the model

y
numeric vector with outputs / scores. If provided then it shall have the same size as data

predict_function
function that takes two arguments: model and new data and returns numeric vector with predictions

residual_function
function that takes three arguments: model, data and response vector y. It should return a numeric vector with model residuals for given data. If not provided, response residuals (\(y-\hat{y}\)) are calculated.

...
other parameters

label
character - the name of the model. By default it's extracted from the 'class' attribute of the model

verbose
if TRUE (default) then diagnostic messages will be printed

precalculate
if TRUE (default) then 'predicted_values' and 'residuals' are calculated when explainer is created. This will happenn also if 'verbose' is TRUE

Value

explainer object ready to work with DALEX

Examples

titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
titanic_train <- read.csv(system.file("extdata", "titanic_train.csv", package = "DALEXtra"))
library("mlr")

#> Loading required package: ParamHelpers

task <- mlr::makeClassifTask(
id = "R",
data = titanic_train,
target = "survived"
)
learner <- mlr::makeLearner(
  "classif.gbm",
  par.vals = list(
    distribution = "bernoulli",
    n.trees = 5000,
    interaction.depth = 4,
    n.minobsinnode = 12,
    shrinkage = 0.001,
    bag.fraction = 0.5,
    train.fraction = 1
  ),
  predict.type = "prob"
)
gbm <- mlr::train(learner, task)
explain_mlr(gbm, titanic_test[,1:17], titanic_test[,18])

#> Preparation of a new explainer is initiated
#>   -> model label       :  WrappedModel  (default)
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  yhat_mlr 
#>   -> predicted values  :  numerical, min =  0.03946409 , mean =  0.3320766 , max =  0.9472849  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.9311745 , mean =  -0.02100794 , max =  0.9502293  
#> A new explainer has been created!

#> Model label:  WrappedModel 
#> Model class:  WrappedModel 
#> Data head  :
#>   gender.female gender.male age class.1st class.2nd class.3rd class.deck.crew
#> 1             1           0  16         0         0         1               0
#> 2             0           1  25         0         0         1               0
#>   class.engineering.crew class.restaurant.staff class.victualling.crew
#> 1                      0                      0                      0
#> 2                      0                      0                      0
#>   embarked.Belfast embarked.Cherbourg embarked.Queenstown embarked.Southampton
#> 1                0                  0                   0                    1
#> 2                0                  0                   0                    1
#>   fare sibsp parch
#> 1 7.13     0     0
#> 2 7.13     0     0