Function for getting binary matrix

Function creates binary matrix, to be used in aspect_importance method. It starts with a zero matrix. Then it replaces some zeros with ones. It either randomly replaces one or two zeros per row. Or replace random number of zeros per row - average number of replaced zeros can be controled by parameter f. Function doesn't allow the returned matrix to have rows with only zeros.

get_sample(n, p, sample_method = c("default", "binom"), f = 2)

Arguments

n
number of rows

p
number of columns

sample_method
sampling method

f
frequency for binomial sampling

Value

a binary matrix

Examples

 # \dontrun{
 get_sample(100,6,"binom",3)

#> Warning: Function is deprecated. Please use triplot package

#>        [,1] [,2] [,3] [,4] [,5] [,6]
#>   [1,]    0    0    1    0    1    0
#>   [2,]    0    1    1    1    0    0
#>   [3,]    1    0    1    1    1    0
#>   [4,]    1    0    0    1    0    1
#>   [5,]    0    0    0    0    1    1
#>   [6,]    0    1    0    0    1    0
#>   [7,]    0    1    1    0    0    1
#>   [8,]    0    0    0    0    0    1
#>   [9,]    0    1    1    0    0    0
#>  [10,]    0    1    0    0    0    1
#>  [11,]    0    1    1    0    1    1
#>  [12,]    0    1    1    0    1    0
#>  [13,]    1    1    1    0    0    1
#>  [14,]    0    0    0    1    0    1
#>  [15,]    0    0    0    1    0    1
#>  [16,]    1    0    1    0    1    0
#>  [17,]    0    1    1    1    0    1
#>  [18,]    1    0    1    1    0    0
#>  [19,]    1    0    0    1    0    0
#>  [20,]    0    1    1    0    0    0
#>  [21,]    1    0    0    0    1    0
#>  [22,]    0    0    0    1    1    0
#>  [23,]    1    1    1    1    0    1
#>  [24,]    0    0    1    1    1    1
#>  [25,]    0    1    0    1    1    0
#>  [26,]    1    1    0    0    1    0
#>  [27,]    0    1    1    0    1    1
#>  [28,]    0    1    1    1    1    0
#>  [29,]    0    0    0    0    1    1
#>  [30,]    0    0    0    0    0    1
#>  [31,]    0    0    1    1    0    1
#>  [32,]    1    0    1    1    1    1
#>  [33,]    1    1    0    0    0    0
#>  [34,]    0    1    0    0    1    0
#>  [35,]    0    1    0    0    1    0
#>  [36,]    1    0    0    1    0    0
#>  [37,]    0    0    0    1    1    0
#>  [38,]    0    1    0    1    1    0
#>  [39,]    0    1    1    0    0    0
#>  [40,]    1    1    0    1    0    1
#>  [41,]    0    1    1    0    1    0
#>  [42,]    1    0    0    1    0    0
#>  [43,]    0    0    1    0    0    0
#>  [44,]    0    1    0    0    0    1
#>  [45,]    0    1    0    1    0    1
#>  [46,]    1    0    1    1    1    0
#>  [47,]    0    0    1    0    1    1
#>  [48,]    1    0    0    0    1    0
#>  [49,]    0    1    0    0    1    0
#>  [50,]    0    1    0    0    0    0
#>  [51,]    1    0    1    0    1    1
#>  [52,]    0    1    1    0    1    1
#>  [53,]    0    1    0    0    1    1
#>  [54,]    0    1    1    0    0    0
#>  [55,]    1    0    1    1    1    1
#>  [56,]    0    0    0    0    1    1
#>  [57,]    0    1    0    1    0    0
#>  [58,]    1    0    0    1    0    1
#>  [59,]    1    1    0    0    1    0
#>  [60,]    0    1    0    0    1    1
#>  [61,]    1    0    0    0    1    0
#>  [62,]    1    0    0    1    0    0
#>  [63,]    0    1    0    1    0    0
#>  [64,]    0    0    0    0    1    1
#>  [65,]    0    0    0    0    1    1
#>  [66,]    0    1    0    1    1    1
#>  [67,]    1    0    0    1    0    0
#>  [68,]    1    0    1    0    1    0
#>  [69,]    1    1    1    0    0    0
#>  [70,]    0    0    0    1    1    1
#>  [71,]    1    1    0    0    1    0
#>  [72,]    0    1    1    0    0    1
#>  [73,]    0    0    1    0    0    0
#>  [74,]    1    0    0    1    0    1
#>  [75,]    1    1    0    0    0    1
#>  [76,]    0    0    1    0    0    0
#>  [77,]    1    0    0    0    1    1
#>  [78,]    1    0    0    0    1    1
#>  [79,]    0    0    0    1    0    1
#>  [80,]    0    0    0    0    1    0
#>  [81,]    1    0    0    0    0    1
#>  [82,]    0    1    1    0    0    0
#>  [83,]    1    0    1    0    0    1
#>  [84,]    0    1    0    1    1    0
#>  [85,]    0    1    0    0    1    0
#>  [86,]    0    0    0    1    1    0
#>  [87,]    0    0    1    1    1    0
#>  [88,]    0    1    1    0    0    1
#>  [89,]    1    1    0    0    1    1
#>  [90,]    1    1    0    0    1    0
#>  [91,]    0    1    0    0    0    1
#>  [92,]    1    1    0    1    0    1
#>  [93,]    0    0    0    1    1    1
#>  [94,]    0    0    1    1    1    0
#>  [95,]    1    1    0    0    1    0
#>  [96,]    1    0    0    0    0    1
#>  [97,]    0    1    0    1    1    0
#>  [98,]    0    0    0    0    0    1
#>  [99,]    0    0    1    0    0    1
#> [100,]    1    0    1    0    1    0

# }