Exract info from model

This generic function let user extract base information about model. The function returns a named list of class model_info that contain about package of model, version and task type. For wrappers like mlr or caret both, package and wrapper information are stored

# S3 method for WrappedModel
model_info(model, is_multiclass = FALSE, ...)

# S3 method for H2ORegressionModel
model_info(model, is_multiclass = FALSE, ...)

# S3 method for H2OBinomialModel
model_info(model, is_multiclass = FALSE, ...)

# S3 method for H2OMultinomialModel
model_info(model, is_multiclass = FALSE, ...)

# S3 method for scikitlearn_model
model_info(model, is_multiclass = FALSE, ...)

# S3 method for keras
model_info(model, is_multiclass = FALSE, ...)

# S3 method for LearnerRegr
model_info(model, is_multiclass = FALSE, ...)

# S3 method for LearnerClassif
model_info(model, is_multiclass = FALSE, ...)

# S3 method for GraphLearner
model_info(model, is_multiclass = FALSE, ...)

# S3 method for xgb.Booster
model_info(model, is_multiclass = FALSE, ...)

# S3 method for workflow
model_info(model, is_multiclass = FALSE, ...)

Arguments

model

- model object

is_multiclass

- if TRUE and task is classification, then multitask classification is set. Else is omitted. If model_info was executed withing explain function. DALEX will recognize subtype on it's own. @param is_multiclass

...

- another arguments

Currently supported packages are:

mlr models created with mlr package

h2o models created with h2o package

scikit-learn models created with scikit-learn Python library and accessed via reticulate

keras models created with keras Python library and accessed via reticulate

mlr3 models created with mlr3 package

xgboost models created with xgboost package

tidymodels models created with tidymodels package

Value

A named list of class model_info