Print funnel_measure object

Print funnel_measure object

# S3 method for training_test_comparison
print(x, ...)

Arguments

x
an object of class funnel_measure

...
other parameters

Examples

# \donttest{
library(DALEXtra)
titanic_train <- read.csv(system.file("extdata", "titanic_train.csv", package = "DALEXtra"))
titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
h2o::h2o.init()

#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         36 seconds 97 milliseconds 
#>     H2O cluster timezone:       UTC 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.30.0.1 
#>     H2O cluster version age:    3 months and 11 days !!! 
#>     H2O cluster name:           H2O_started_from_R_runner_ssh825 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   1.64 GB 
#>     H2O cluster total cores:    1 
#>     H2O cluster allowed cores:  1 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, XGBoost, Algos, AutoML, Core V3, TargetEncoder, Core V4 
#>     R Version:                  R version 4.0.2 (2020-06-22) 

#> Warning: 
#> Your H2O cluster version is too old (3 months and 11 days)!
#> Please download and install the latest version from http://h2o.ai/download/

#> 

h2o::h2o.no_progress()
titanic_h2o <- h2o::as.h2o(titanic_train)

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

titanic_h2o["survived"] <- h2o::as.factor(titanic_h2o["survived"])
titanic_test_h2o <- h2o::as.h2o(titanic_test)

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

model <- h2o::h2o.gbm(
  training_frame = titanic_h2o,
  y = "survived",
  distribution = "bernoulli",
  ntrees = 500,
  max_depth = 4,
  min_rows =  12,
  learn_rate = 0.001
)
explainer_h2o <- explain_h2o(model, titanic_test[,1:17], titanic_test[,18])

#> Preparation of a new explainer is initiated
#>   -> model label       :  H2OBinomialModel  (  default  )
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  yhat.H2OBinomialModel  will be used (  default  )

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   -> predicted values  :  numerical, min =  0.1994385 , mean =  0.3299534 , max =  0.5876636  

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   -> model_info        :  package h2o , ver. 3.30.0.1 , task classification (  default  ) 
#>   -> residual function :  difference between y and yhat (  default  )

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   -> residuals         :  numerical, min =  -0.5847857 , mean =  -0.01888466 , max =  0.8005615  
#>   A new explainer has been created!  

explainer_scikit <- explain_scikitlearn(system.file("extdata",
                                                    "scikitlearn.pkl",
                                                    package = "DALEXtra"),
                                        yml = system.file("extdata",
                                                          "testing_environment.yml",
                                                          package = "DALEXtra"),
                                        data = titanic_test[,1:17],
                                        y = titanic_test$survived)

#> Path to conda not specified while on unix. Default used. /usr/share/miniconda/bin/..

#> There already exists environment named the same as it is specified in .yml file - myenv. It will be used

#> Preparation of a new explainer is initiated
#>   -> model label       :  scikitlearn_model  (  default  )
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  yhat.scikitlearn_model  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.02086126 , mean =  0.288584 , max =  0.9119996  
#>   -> model_info        :  package reticulate , ver. 1.16 , task classification (  default  ) 
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.8669431 , mean =  0.02248468 , max =  0.9791387  
#>   A new explainer has been created!  

library("mlr")
task <- mlr::makeClassifTask(
  id = "R",
  data = titanic_train,
  target = "survived"
)
learner <- mlr::makeLearner(
  "classif.gbm",
  par.vals = list(
    distribution = "bernoulli",
    n.trees = 500,
    interaction.depth = 4,
    n.minobsinnode = 12,
    shrinkage = 0.001,
    bag.fraction = 0.5,
    train.fraction = 1
  ),
  predict.type = "prob"
)
gbm <- mlr::train(learner, task)
explainer_mlr <- explain_mlr(gbm, titanic_test[,1:17], titanic_test[,18])

#> Preparation of a new explainer is initiated
#>   -> model label       :  WrappedModel  (  default  )
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  yhat.WrappedModel  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.4352853 , mean =  0.6677554 , max =  0.7298008  
#>   -> model_info        :  package mlr , ver. 2.17.1 , task classification (  default  ) 
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7298008 , mean =  -0.3566867 , max =  0.5647147  
#>   A new explainer has been created!  

data <- training_test_comparison(explainer_scikit, list(explainer_h2o, explainer_mlr),
                                 training_data = titanic_train[,-18],
                                 training_y = titanic_train[,18])

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

print(data)

#> Training test data head:
#>   measure_test measure_train             label       type
#> 1    0.2085720     0.2274120  H2OBinomialModel Challenger
#> 2    0.8115324     0.8120089      WrappedModel Challenger
#> 3    0.1915266     0.1603138 scikitlearn_model   Champion
#> Models Info
#>               label
#> 1 scikitlearn_model
#> 2  H2OBinomialModel
#> 3      WrappedModel
#>                                                           class       type
#> 1 sklearn.ensemble.gradient_boosting.GradientBoostingClassifier   Champion
#> 2                                              H2OBinomialModel Challenger
#> 3                                                  WrappedModel Challenger

# }