Wrapper for Python Scikit-Learn Models

scikit-learn models may be loaded into R environment like any other Python object. This function helps to inspect performance of Python model and compare it with other models, using R tools like DALEX. This function creates an object that is easy accessible R version of scikit-learn model exported from Python via pickle file.

scikitlearn_explain(path)

Arguments

path
a path to the pickle file. Can be used without other arguments if you are sure that active Python version match pickle version.

yml
a path to the yml file. Conda virtual env will be recreated from this file. If OS is Windows conda has to be added to the PATH first

condaenv
If yml param is provided, a path to the main conda folder. If yml is null, a name of existing conda environment.

env
A path to python virtual environment.

data
test data set that will be passed to explainer.

y
vector that will be passed to explainer.

predict_function
predict function that will be passed into explainer. If NULL, default will be used.

residual_function
residual function that will be passed into explainer. If NULL, default will be used.

label
label that will be passed into explainer. If NULL, default will be used.

verbose
bool that will be passed into explainer. If NULL, default will be used.

Value

An object of the class 'explainer'. Possible extraction of object 'scikitlearn_model' via explainer$model

scikitlearn_model is a list with following fields:

model it is original model received vie reticiulate function. Use it for computations.

predict_function predict function extracted from original model. It is adjusted to DALEX demands and therfore fully compatibile.

type type model, classification or regression

params object of class `scikitlearn_set` which in fact is list that consist of parameters of our model.

label name of model

Example of Python code

from pandas import DataFrame, read_csv
import pandas as pd
import pickle
import sklearn.ensemble
model = sklearn.ensemble.GradientBoostingClassifier()
model = model.fit(titanic_train_X, titanic_train_Y)
pickle.dump(model, open("gbm.pkl", "wb"), protocol = 2)

In order to export environment into .yml, activating virtual env via activate name_of_the_env and execution of the following shell command is necessary
conda env export > environment.yml

Errors use case
Here is shortened version of solution for specific errors

There already exists environment with a name specified by given .yml file
You have two ways of solving that issue. Both connected with anaconda prompt. First is removing conda env with command:
conda env remove --name myenv
And execute function once again. Second is updating env via:
conda env create -f environment.yml

Conda cannot find specified packages at channels you have provided.
That error may be casued by a lot of things. Of of those is that specified version is too old to be avaialble from offcial conda repo. Edit Your .yml file and add link to proper repository at channels section.

Issue may be also connected with the platform. If model was created on the platform with different OS yo may need to remove specific version from .yml file.
- numpy=1.16.4=py36h19fb1c0_0
- numpy-base=1.16.4=py36hc3f5095_0
In the example above You have to remove =py36h19fb1c0_0 and =py36hc3f5095_0
If some packages are not availbe for anaconda at all, use pip statement

If .yml file seems not to work, virtual env can be created manually using anaconda promt.
conda create -n name_of_env python=3.4
conda install -n name_of_env name_of_package=0.20

Examples

# Usage with explain()
reticulate::use_condaenv("myenv")
have_sklearn <- reticulate::py_module_available("sklearn.ensemble")
library("DALEX")

#> Welcome to DALEX (version: 0.4.5).
#> Find examples and detailed introduction at: https://pbiecek.github.io/PM_VEE/

library("DALEXtra")
library("reticulate")

if(have_sklearn) {
   # Explainer build (Keep in mind that 18th column is target)
   titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
   # Keep in mind that when pickle is being built and loaded,
   # not only Python version but libraries versions has to match aswell
   model <- scikitlearn_explain(system.file("extdata", "scikitlearn.pkl", package = "DALEXtra"), condaenv = "myenv")
   explainer <- explain(model = model, data = titanic_test[,1:17], y = titanic_test$survived, predict_function = model$predict_function)
   print(model_performance(explainer))

   # Predictions with newdata
   predictions <- model$predict_function(model, titanic_test[,1:17])

} else {
  print('Python testing environment is required.')
}

#> Error in scikitlearn_explain(system.file("extdata", "scikitlearn.pkl",     package = "DALEXtra"), condaenv = "myenv"): object 'explainer' not found