Compare machine learning models

Determining if one model is better than the other one is a difficult task. Mostly because there is a lot of fields that have to be covered to make such a judgemnt. Overall performance, performance on the crucial subset, distribution of residuals, those are only few among many ideas related to that issue. Following function allow user to create a report based on various sections. Each says something different about relation between champion and challengers. DALEXtra package share 3 base sections which are funnel_measure overall_comparison and training_test_comparison but any object that has generic plot function can be inculded at report.

champion_challenger(sections, dot_size = 4, output_dir_path = getwd(),
  output_name = "Report", model_performance_table = FALSE,
  title = "ChampionChallenger", author = Sys.info()[["user"]])

Arguments

sections
- list of sections to be attached to report. Could be sections available with DALEXtra which are funnel_measure training_test_comparison, overall_comparison or any other explanation that can work with plot function. Please provide name for not standard sections, that will be presented as section titles. Oterwise class of the object will be used.

dot_size
- dot_size argument passed to plot.funnel_measure if funnel_measure section present

output_dir_path
- path to directory where Report should be created. By default it is current working directory.

output_name
- name of the Report. By default it is "Report"

model_performance_table
- If TRUE and overall_comparison section present, table of scores will be displayed.

title
- Title for report, by default it is "ChampionChallenger".

author
- Author of , report. By default it is current user name.

Value

rmarkdown report

Examples

library("mlr")

#> Ładowanie wymaganego pakietu: ParamHelpers

library("DALEXtra")
task <- mlr::makeRegrTask(
  id = "R",
  data = apartments,
  target = "m2.price"
)
learner_lm <- mlr::makeLearner(
  "regr.lm"
)
model_lm <- mlr::train(learner_lm, task)
explainer_lm <- explain_mlr(model_lm, apartmentsTest, apartmentsTest$m2.price, label = "LM")

#> Preparation of a new explainer is initiated
#>   -> model label       :  LM 
#>   -> data              :  9000  rows  6  cols 
#>   -> target variable   :  9000  values 
#>   -> data              :  A column identical to the target variable `y` has been found in the `data`.  (WARNING)
#>   -> data              :  It is highly recommended to pass `data` without the target variable column
#>   -> predict function  :  yhat.WrappedModel  will be used (default)
#>   -> predicted values  :  numerical, min =  1792.597 , mean =  3506.836 , max =  6241.447  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -257.2555 , mean =  4.687686 , max =  472.356  
#> A new explainer has been created!


learner_rf <- mlr::makeLearner(
  "regr.randomForest"
)
model_rf <- mlr::train(learner_rf, task)
explainer_rf <- explain_mlr(model_rf, apartmentsTest, apartmentsTest$m2.price, label = "RF")

#> Preparation of a new explainer is initiated
#>   -> model label       :  RF 
#>   -> data              :  9000  rows  6  cols 
#>   -> target variable   :  9000  values 
#>   -> data              :  A column identical to the target variable `y` has been found in the `data`.  (WARNING)
#>   -> data              :  It is highly recommended to pass `data` without the target variable column
#>   -> predict function  :  yhat.WrappedModel  will be used (default)
#>   -> predicted values  :  numerical, min =  1976.589 , mean =  3507.215 , max =  5779.887  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -715.967 , mean =  4.308321 , max =  1255.235  
#> A new explainer has been created!


learner_gbm <- mlr::makeLearner(
  "regr.gbm"
)
model_gbm <- mlr::train(learner_gbm, task)
explainer_gbm <- explain_mlr(model_gbm, apartmentsTest, apartmentsTest$m2.price, label = "GBM")

#> Preparation of a new explainer is initiated
#>   -> model label       :  GBM 
#>   -> data              :  9000  rows  6  cols 
#>   -> target variable   :  9000  values 
#>   -> data              :  A column identical to the target variable `y` has been found in the `data`.  (WARNING)
#>   -> data              :  It is highly recommended to pass `data` without the target variable column
#>   -> predict function  :  yhat.WrappedModel  will be used (default)
#>   -> predicted values  :  numerical, min =  2129.738 , mean =  3502.03 , max =  6067.672  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -628.0318 , mean =  9.493952 , max =  732.9962  
#> A new explainer has been created!



plot_data <- funnel_measure(explainer_lm, list(explainer_rf, explainer_gbm),
                            nbins = 5, measure_function = DALEX::loss_root_mean_square)

#>   |                                                                              |                                                                      |   0%  |                                                                              |============                                                          |  17%  |                                                                              |=======================                                               |  33%  |                                                                              |===================================                                   |  50%  |                                                                              |===============================================                       |  67%  |                                                                              |==========================================================            |  83%  |                                                                              |======================================================================| 100%


champion_challenger(list(plot_data), dot_size = 3)

#> 
#> 
#> processing file: Report.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |........                                                         |  12%
#>    inline R code fragments
#> 
#>   |                                                                         |................                                                 |  25%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                         |........................                                         |  38%
#>   ordinary text without R code
#> 
#>   |                                                                         |................................                                 |  50%
#> label: unnamed-chunk-1
#>   |                                                                         |.........................................                        |  62%
#>   ordinary text without R code
#> 
#>   |                                                                         |.................................................                |  75%
#> label: unnamed-chunk-2 (with options) 
#> List of 2
#>  $ fig.width : num 10
#>  $ fig.height: num 8
#> 

#>   |                                                                         |.........................................................        |  88%
#>   ordinary text without R code
#> 
#>   |                                                                         |.................................................................| 100%
#> label: unnamed-chunk-3
#> 

#> output file: Report.knit.md

#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS Report.utf8.md --to html4 --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output pandoc49c6d39289d.html --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_collapsed=1 --variable toc_smooth_scroll=1 --variable toc_print=1 --template "C:\Users\szyna\Documents\R\win-library\3.6\rmarkdown\rmd\h\default.html" --no-highlight --variable highlightjs=1 --number-sections --variable "theme:bootstrap" --include-in-header "C:\Users\szyna\AppData\Local\Temp\RtmpMRBUBd\rmarkdown-str49cfe27a07.html" --mathjax --variable "mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" 

#> 
#> Output created: Report.html