Create your conda virtual env with DALEX

Python objects may be loaded into R. However, it requiers versions of the Python and libraries to match between both machines. This functions allow user to create conda virtual environment based on provided .yml file.

create_env(yml, condaenv)

Arguments

yml
a path to the .yml file. If OS is Windows conda has to be added to the PATH first

condaenv
path to main conda folder. If OS is Unix You may want to specify it. When passed with windows, param will be omitted.

Value

Name of created virtual env.

Examples

if(DALEXtra:::is_conda()) {
  create_env(system.file("extdata", "testing_environment.yml", package = "DALEXtra"))
} else {
  "conda is required"
}

#> There already exists environment named the same as it is specified in .yml file - myenv. It will be used

#> [1] "myenv"