Create explainer from your h2o model

DALEX is designed to work with various black-box models like tree ensembles, linear models, neural networks etc. Unfortunately R packages that create such models are very inconsistent. Different tools use different interfaces to train, validate and use models. One of those tools, we would like to make more accessible is H2O.

explain_h2o(model, data = NULL, y = NULL, predict_function = NULL,
  residual_function = NULL, ..., label = NULL, verbose = TRUE,
  precalculate = TRUE)

Arguments

model
object - a model to be explained

data
data.frame or matrix - data that was used for fitting. If not provided then will be extracted from the model. Data should be passed without target column (this shall be provided as the y argument). NOTE: If target variable is present in the data, some of the functionalities my not work properly.

y
numeric vector with outputs / scores. If provided then it shall have the same size as data

predict_function
function that takes two arguments: model and new data and returns numeric vector with predictions

residual_function
function that takes three arguments: model, data and response vector y. It should return a numeric vector with model residuals for given data. If not provided, response residuals (\(y-\hat{y}\)) are calculated.

...
other parameters (passed for example to predict function)

label
character - the name of the model. By default it's extracted from the 'class' attribute of the model

verbose
if TRUE (default) then diagnostic messages will be printed

precalculate
if TRUE (default) then 'predicted_values' and 'residuals' are calculated when explainer is created. This will happenn also if 'verbose' is TRUE.

Value

explainer object (explain) ready to work with DALEX

Examples

library("DALEXtra")
titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
titanic_train <- read.csv(system.file("extdata", "titanic_train.csv", package = "DALEXtra"))
h2o::h2o.init()

#> 
#> H2O is not running yet, starting it now...
#> 
#> Note:  In case of errors look at the following log files:
#>     C:\Users\szyna\AppData\Local\Temp\RtmpMbVK2r/h2o_szyna_started_from_r.out
#>     C:\Users\szyna\AppData\Local\Temp\RtmpMbVK2r/h2o_szyna_started_from_r.err
#> 
#> 
#> Starting H2O JVM and connecting: . Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 seconds 643 milliseconds 
#>     H2O cluster timezone:       Europe/Belgrade 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.24.0.5 
#>     H2O cluster version age:    3 months and 26 days !!! 
#>     H2O cluster name:           H2O_started_from_R_szyna_ozw432 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   1.97 GB 
#>     H2O cluster total cores:    4 
#>     H2O cluster allowed cores:  4 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.6.1 (2019-07-05) 

#> Warning: 
#> Your H2O cluster version is too old (3 months and 26 days)!
#> Please download and install the latest version from http://h2o.ai/download/

#> 

h2o::h2o.no_progress()
titanic_h2o <- h2o::as.h2o(titanic_train)
titanic_h2o["survived"] <- h2o::as.factor(titanic_h2o["survived"])
titanic_test_h2o <- h2o::as.h2o(titanic_test)
model <- h2o::h2o.gbm(
training_frame = titanic_h2o,
y = "survived",
distribution = "bernoulli",
ntrees = 500,
max_depth = 4,
min_rows =  12,
learn_rate = 0.001
)
explain_h2o(model, titanic_test[,1:17], titanic_test[,18])

#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         11 seconds 734 milliseconds 
#>     H2O cluster timezone:       Europe/Belgrade 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.24.0.5 
#>     H2O cluster version age:    3 months and 26 days !!! 
#>     H2O cluster name:           H2O_started_from_R_szyna_ozw432 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   1.97 GB 
#>     H2O cluster total cores:    4 
#>     H2O cluster allowed cores:  4 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.6.1 (2019-07-05) 

#> Warning: 
#> Your H2O cluster version is too old (3 months and 26 days)!
#> Please download and install the latest version from http://h2o.ai/download/

#> 
#> Preparation of a new explainer is initiated
#>   -> model label       :  H2OBinomialModel  (  default  )
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  yhat.H2OBinomialModel  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.1994385 , mean =  0.3299534 , max =  0.5876636  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.5847857 , mean =  -0.01888466 , max =  0.8005615  
#>   -> model_info        :  model_info not specified, assuming type: regression (  default  ) 
#>   -> model_info        :  package: Model of class: H2OBinomialModel package unrecognized Unknown (  default  ) 
#>   A new explainer has been created!  

#> Model label:  H2OBinomialModel 
#> Model class:   
#> Data head  :
#>   gender.female gender.male age class.1st class.2nd class.3rd class.deck.crew
#> 1             1           0  16         0         0         1               0
#> 2             0           1  25         0         0         1               0
#>   class.engineering.crew class.restaurant.staff class.victualling.crew
#> 1                      0                      0                      0
#> 2                      0                      0                      0
#>   embarked.Belfast embarked.Cherbourg embarked.Queenstown embarked.Southampton
#> 1                0                  0                   0                    1
#> 2                0                  0                   0                    1
#>   fare sibsp parch
#> 1 7.13     0     0
#> 2 7.13     0     0

h2o::h2o.shutdown(prompt = FALSE)

#> [1] TRUE