Create explainer from your h2o model

DALEX is designed to work with various black-box models like tree ensembles, linear models, neural networks etc. Unfortunately R packages that create such models are very inconsistent. Different tools use different interfaces to train, validate and use models. One of those tools, we would like to make more accessible is H2O.

explain_h2o(
  model,
  data = NULL,
  y = NULL,
  weights = NULL,
  predict_function = NULL,
  residual_function = NULL,
  ...,
  label = NULL,
  verbose = TRUE,
  precalculate = TRUE,
  colorize = TRUE,
  model_info = NULL,
  type = NULL
)

Arguments

model
object - a model to be explained

data
data.frame or matrix - data that was used for fitting. If not provided then will be extracted from the model. Data should be passed without target column (this shall be provided as the y argument). NOTE: If target variable is present in the data, some of the functionalities my not work properly.

y
numeric vector with outputs / scores. If provided then it shall have the same size as data

weights
numeric vector with sampling weights. By default it's NULL. If provided then it shall have the same length as data

predict_function
function that takes two arguments: model and new data and returns numeric vector with predictions

residual_function
function that takes three arguments: model, data and response vector y. It should return a numeric vector with model residuals for given data. If not provided, response residuals (\(y-\hat{y}\)) are calculated.

...
other parameters

label
character - the name of the model. By default it's extracted from the 'class' attribute of the model

verbose
if TRUE (default) then diagnostic messages will be printed

precalculate
if TRUE (default) then 'predicted_values' and 'residuals' are calculated when explainer is created.

colorize
if TRUE (default) then WARNINGS, ERRORS and NOTES are colorized. Will work only in the R console.

model_info
a named list (package, version, type) containg information about model. If NULL, DALEX will seek for information on it's own.

type
type of a model, either classification or regression. If not specified then type will be extracted from model_info.

Value

explainer object (explain) ready to work with DALEX

Examples

# \donttest{

# # load packages and data
library(h2o)

#> 
#> ----------------------------------------------------------------------
#> 
#> Your next step is to start H2O:
#>     > h2o.init()
#> 
#> For H2O package documentation, ask for help:
#>     > ??h2o
#> 
#> After starting H2O, you can use the Web UI at http://localhost:54321
#> For more information visit http://docs.h2o.ai
#> 
#> ----------------------------------------------------------------------

#> 
#> Attaching package: ‘h2o’

#> The following objects are masked from ‘package:stats’:
#> 
#>     cor, sd, var

#> The following objects are masked from ‘package:base’:
#> 
#>     &&, %*%, %in%, ||, apply, as.factor, as.numeric, colnames,
#>     colnames<-, ifelse, is.character, is.factor, is.numeric, log,
#>     log10, log1p, log2, round, signif, trunc

library(DALEXtra)

# data <- DALEX::titanic_imputed

# init h2o
 h2o.init()

#> 
#> H2O is not running yet, starting it now...
#> 
#> Note:  In case of errors look at the following log files:
#>     /tmp/RtmpCQT159/file70db53629856/h2o_runner_started_from_r.out
#>     /tmp/RtmpCQT159/file70db4d52c259/h2o_runner_started_from_r.err
#> 
#> 
#> Starting H2O JVM and connecting: ... Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         2 seconds 875 milliseconds 
#>     H2O cluster timezone:       UTC 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.30.0.1 
#>     H2O cluster version age:    4 months and 17 days !!! 
#>     H2O cluster name:           H2O_started_from_R_runner_tjh847 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   1.51 GB 
#>     H2O cluster total cores:    2 
#>     H2O cluster allowed cores:  2 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, XGBoost, Algos, AutoML, Core V3, TargetEncoder, Core V4 
#>     R Version:                  R version 4.0.2 (2020-06-22) 

#> Warning: 
#> Your H2O cluster version is too old (4 months and 17 days)!
#> Please download and install the latest version from http://h2o.ai/download/

#> 

# stop h2o progress printing
 h2o.no_progress()

# split the data
# h2o_split <- h2o.splitFrame(as.h2o(data))
# train <- h2o_split[[1]]
# test <- as.data.frame(h2o_split[[2]])
# h2o automl takes target as factor
# train$survived <- as.factor(train$survived)

# fit a model
# automl <- h2o.automl(y = "survived",
#                   training_frame = train,
#                    max_runtime_secs = 30)

# create an explainer for the model
# explainer <- explain_h2o(automl,
#                        data = test,
#                         y = test$survived,
#                          label = "h2o")

titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
titanic_train <- read.csv(system.file("extdata", "titanic_train.csv", package = "DALEXtra"))
titanic_h2o <- h2o::as.h2o(titanic_train)

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

titanic_h2o["survived"] <- h2o::as.factor(titanic_h2o["survived"])
titanic_test_h2o <- h2o::as.h2o(titanic_test)

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

model <- h2o::h2o.gbm(
training_frame = titanic_h2o,
y = "survived",
distribution = "bernoulli",
ntrees = 500,
max_depth = 4,
min_rows =  12,
learn_rate = 0.001
)
explain_h2o(model, titanic_test[,1:17], titanic_test[,18])

#> Preparation of a new explainer is initiated
#>   -> model label       :  H2OBinomialModel  (  default  )
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> predict function  :  yhat.H2OBinomialModel  will be used (  default  )

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   -> predicted values  :  numerical, min =  0.1994385 , mean =  0.3299534 , max =  0.5876636  

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   -> model_info        :  package h2o , ver. 3.30.0.1 , task classification (  default  ) 
#>   -> residual function :  difference between y and yhat (  default  )

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   -> residuals         :  numerical, min =  -0.5847857 , mean =  -0.01888466 , max =  0.8005615  
#>   A new explainer has been created!  

#> Model label:  H2OBinomialModel 
#> Model class:  H2OBinomialModel 
#> Data head  :
#>   gender.female gender.male age class.1st class.2nd class.3rd class.deck.crew
#> 1             1           0  16         0         0         1               0
#> 2             0           1  25         0         0         1               0
#>   class.engineering.crew class.restaurant.staff class.victualling.crew
#> 1                      0                      0                      0
#> 2                      0                      0                      0
#>   embarked.Belfast embarked.Cherbourg embarked.Queenstown embarked.Southampton
#> 1                0                  0                   0                    1
#> 2                0                  0                   0                    1
#>   fare sibsp parch
#> 1 7.13     0     0
#> 2 7.13     0     0

h2o.shutdown(prompt = FALSE)
# }