Function for getting binary matrix

Function creates binary matrix, to be used in aspect_importance method. It starts with a zero matrix. Then it replaces some zeros with ones. It either randomly replaces one or two zeros per row. Or replace random number of zeros per row - average number of replaced zeros can be controled by parameter f. Function doesn't allow the returned matrix to have rows with only zeros.

get_sample(n, p, sample_method = c("default", "binom"), f = 2)

Arguments

n
number of rows

p
number of columns

sample_method
sampling method

f
frequency for binomial sampling

Value

a binary matrix

Examples

 
# NOT RUN {
 get_sample(100,6,"binom",3)
# }