Compare performance of model between training and test set

Function training_test_comparison calculates performance of the provided model based on specified measure function. Response of the model is caluclated based on test data, extracted from the explainer and training data, provided by the user. Output can be easily shown with print or plot function.

training_test_comparison(
  champion,
  challengers,
  training_data,
  training_y,
  measure_function = DALEX::loss_root_mean_square
)

Arguments

champion
- explainer of champion model.

challengers
- explainer of challenger model or list of explainers.

training_data
- data without target column that will be passed to predict function and then to measure function. Keep in mind that they have to differ from data passed to an explainer.

training_y
- target column for training_data

measure_function
- measure function that calculates performance of model based on true observation and prediction. Order of parameters is important and should be (y, y_hat). By default it is RMSE.

Value

An object of the class training_test_comparison.

It is a named list containig:

data data.frame with following columns

measure_test performance on test set

measure_train performance on training set

label label of explainer

type flag that indicates if explainer was passed as champion or as challenger.

models_info data.frame containig inforamtion about models used in analysys

Examples

# \donttest{
library(DALEXtra)
titanic_train <- read.csv(system.file("extdata", "titanic_train.csv", package = "DALEXtra"))
titanic_test <- read.csv(system.file("extdata", "titanic_test.csv", package = "DALEXtra"))
h2o::h2o.init()

#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         34 seconds 332 milliseconds 
#>     H2O cluster timezone:       UTC 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.30.0.1 
#>     H2O cluster version age:    18 days  
#>     H2O cluster name:           H2O_started_from_R_runner_ssh825 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   2.66 GB 
#>     H2O cluster total cores:    4 
#>     H2O cluster allowed cores:  4 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, XGBoost, Algos, AutoML, Core V3, TargetEncoder, Core V4 
#>     R Version:                  R version 3.6.3 (2020-02-29) 
#> 

h2o::h2o.no_progress()
titanic_h2o <- h2o::as.h2o(titanic_train)

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

titanic_h2o["survived"] <- h2o::as.factor(titanic_h2o["survived"])
titanic_test_h2o <- h2o::as.h2o(titanic_test)

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

model <- h2o::h2o.gbm(
  training_frame = titanic_h2o,
  y = "survived",
  distribution = "bernoulli",
  ntrees = 500,
  max_depth = 4,
  min_rows =  12,
  learn_rate = 0.001
)
explainer_h2o <- explain_h2o(model, titanic_test[,1:17], titanic_test[,18])

#> Preparation of a new explainer is initiated
#>   -> model label       :  H2OBinomialModel  (  default  )
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> model_info        :  package h2o , ver. 3.30.0.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.H2OBinomialModel  will be used (  default  )

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   -> predicted values  :  numerical, min =  0.4123364 , mean =  0.6700466 , max =  0.8005615  
#>   -> residual function :  difference between y and yhat (  default  )

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   -> residuals         :  numerical, min =  -0.8005615 , mean =  -0.3589779 , max =  0.5876636  
#>   A new explainer has been created!  

explainer_scikit <- explain_scikitlearn(system.file("extdata",
                                                    "scikitlearn.pkl",
                                                    package = "DALEXtra"),
                                        yml = system.file("extdata",
                                                          "testing_environment.yml",
                                                          package = "DALEXtra"),
                                        data = titanic_test[,1:17],
                                        y = titanic_test$survived)

#> Error in create_env(yml, condaenv): Conda not found

library("mlr")
task <- mlr::makeClassifTask(
  id = "R",
  data = titanic_train,
  target = "survived"
)
learner <- mlr::makeLearner(
  "classif.gbm",
  par.vals = list(
    distribution = "bernoulli",
    n.trees = 500,
    interaction.depth = 4,
    n.minobsinnode = 12,
    shrinkage = 0.001,
    bag.fraction = 0.5,
    train.fraction = 1
  ),
  predict.type = "prob"
)
gbm <- mlr::train(learner, task)
explainer_mlr <- explain_mlr(gbm, titanic_test[,1:17], titanic_test[,18])

#> Preparation of a new explainer is initiated
#>   -> model label       :  WrappedModel  (  default  )
#>   -> data              :  524  rows  17  cols 
#>   -> target variable   :  524  values 
#>   -> model_info        :  package mlr , ver. 2.17.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.WrappedModel  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.4350702 , mean =  0.6678622 , max =  0.7297494  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7297494 , mean =  -0.3567935 , max =  0.5649298  
#>   A new explainer has been created!  

data <- training_test_comparison(explainer_scikit, list(explainer_h2o, explainer_mlr),
                                 training_data = titanic_train[,-18],
                                 training_y = titanic_train[,18])

#> Error in training_test_comparison(explainer_scikit, list(explainer_h2o,     explainer_mlr), training_data = titanic_train[, -18], training_y = titanic_train[,     18]): object 'explainer_scikit' not found

plot(data)

#> Warning: data set ‘x’ not found

#> Error in curve(expr = x, from = from, to = to, xlim = xlim, ylab = ylab,     ...): 'expr' did not evaluate to an object of length 'n'

# }